replace 3 implementation of mock spans with a new MockSpan
