+++ tracer/test/Datadog.Trace.Tests/Tagging/TagsListTests.cs
@@ -221,45 +221,5 @@ namespace Datadog.Trace.Tests.Tagging
                                .BeTrue($"Property {propertyAndTag.property.Name} of type {type.Name} is not mapped");
             }
         }
-
-        [global::MessagePack.MessagePackObject]
-        public struct FakeSpan
-        {
-            [global::MessagePack.Key("trace_id")]
-            public ulong TraceId { get; set; }
-
-            [global::MessagePack.Key("span_id")]
-            public ulong SpanId { get; set; }
-
-            [global::MessagePack.Key("name")]
-            public string Name { get; set; }
-
-            [global::MessagePack.Key("resource")]
-            public string Resource { get; set; }
-
-            [global::MessagePack.Key("service")]
-            public string Service { get; set; }
-
-            [global::MessagePack.Key("type")]
-            public string Type { get; set; }
-
-            [global::MessagePack.Key("start")]
-            public long Start { get; set; }
-
-            [global::MessagePack.Key("duration")]
-            public long Duration { get; set; }
-
-            [global::MessagePack.Key("parent_id")]
-            public ulong? ParentId { get; set; }
-
-            [global::MessagePack.Key("error")]
-            public byte Error { get; set; }
-
-            [global::MessagePack.Key("meta")]
-            public Dictionary<string, string> Tags { get; set; }
-
-            [global::MessagePack.Key("metrics")]
-            public Dictionary<string, double> Metrics { get; set; }
-        }
     }
 }
